Fix missing cast.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@985 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ mplayer.c
@@ -573,7 +573,7 @@ if(!filename){
     printf("Invalid audio output driver name: %s\nUse '-ao help' to get a list of available audio drivers.\n",audio_driver);
     return 0;
   }
-  if(dsp) audio_out->control(AOCONTROL_SET_DEVICE,dsp);
+  if(dsp) audio_out->control(AOCONTROL_SET_DEVICE,(int)dsp);
 
 // check codec.conf
 if(!parse_codec_cfg(get_path("codecs.conf"))){
